refactor(mentor) : 로깅 , validation 의존성 주입
