+++ .gitignore
@@ -5,6 +5,8 @@
 .metadata
 bin/
 tmp/
+log/
+*.log
 *.tmp
 *.bak
 *.swp
@@ -13,6 +15,7 @@ local.properties
 .settings/
 .loadpath
 .recommenders
+.github/
 
 # External tool builders
 .externalToolBuilders/
