+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -5,13 +5,16 @@ import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.UserService;
 import com.sprint.mission.discodeit.validation.UserValidator;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.util.*;
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class BasicUserService implements UserService {
 
     private final UserRepository userRepository;
@@ -22,7 +25,7 @@ public class BasicUserService implements UserService {
         if (userValidator.isValidName(name) && userValidator.isValidEmail(email)) {
             User newUser = User.createUser(name, email, password);
             userRepository.save(newUser);
-            System.out.println("create user: " + newUser.getName());
+            log.info("Create User: {}", newUser);
             return newUser;
         }
         return null;
@@ -36,7 +39,7 @@ public class BasicUserService implements UserService {
     @Override
     public User searchById(UUID id) {
         return userRepository.findById(id) // Optional 을 받아서 처리
-                .orElseThrow(() -> new NoSuchElementException("user does not exist"));
+                .orElseThrow(() -> new NoSuchElementException("User does not exist"));
     }
 
     @Override
@@ -45,7 +48,7 @@ public class BasicUserService implements UserService {
         if (userValidator.isValidName(newName) && userValidator.isValidEmail(newEmail)) {
             user.update(newName, newEmail, newPassword);
             userRepository.save(user);
-            System.out.println("success updateUser");
+            log.info("Update User :{}", user);
         }
     }
