+++ src/main/resources/application.yml
@@ -4,3 +4,9 @@ spring:
 
 server:
   port: 8080
+
+logging:
+#  file:
+#    name: ${user.dir}/log/test.log
+  level:
+    com.sprint.mission.discodeit : info
